split gens file
