+++ annet/generators/exceptions.py
@@ -1,0 +1,10 @@
+class GeneratorError(Exception):
+    pass
+
+
+class InvalidValueFromGenerator(ValueError):
+    pass
+
+
+class NotSupportedDevice(GeneratorError):
+    pass
